az-doc-query/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sentence-transformers==3.1.0
faiss-cpu==1.8.0
python-multipart==0.0.9
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
tiktoken==0.7.0
